New app changes:
Fixed crash entering settings screen on iOS 7.1

Fixed layout of Share button on Run screen.

Fixed text colours for action sheets, as states are toggled (esp. for OFCA app).

Fixed issue where the central text in the dial on the Run screen was too big on iPad.

Results page – removed the passive metrics from here (when you click on an individual result) and instead put the IP and submission ID.

Re-label these so Submission ID says “Reference No.” and Public IP says “Your IP”

Fixed issue where in the History screen, when you click the first entry you see an ‘export’ feature, when you clicked onto the other red bars (the second, third, fourth etc) that isn’t in there.
Note that the Export button is visible on when running a test on 3G/4G (i.e. not on WiFi)